+++ packages/zero/src/behaviors/core.ts
@@ -17,7 +17,7 @@
  */
 
 export type { IdGenerator } from './create-id.js';
-export { createId, useIdGenerator, zeroPlugin } from './create-id.js';
+export { createId, idToken, useIdGenerator, zeroPlugin } from './create-id.js';
 
 export type { ControllableState, ControllableOptions } from './controllable.js';
 export { createControllableState, createInertState, namedModel } from './controllable.js';